#184 - Refactor selects.py - Create selects subdir with __init__ - move selects into their own files within subfolder - update the various files that relied on those select classes
